Updating TurningEncoder from Lamprey
The SparkMax currently uses the Neo encoder to run it's position loop.

When the robot boots up it has no idea where it starts.

The lamprey encoder can read the current location of the azimuth no matter where it boots up.

Added the functionality to set a new value to the Neo encoder based on the lamprey encoder.

Also added an offset that is stored by a FluidConstant and will automatically update the neo motor when the entry is updated.

Also added a command to update the turning encoder from lamprey and made it automatically schedule when the robot boots up.