Modified FluidConstant
I promise I didn't make any breaking changes.

All I did was add the possibility to select whether or not the NetworkTable entry can be changed while enabled.

I made sure that if this version of the FluidConstant ran on previous year's code it would still work as it would only run while disabled and has the same constructor parameters.