Modified FluidConstant
Made another very small change to FluidConstant. This could have repercussions if used with old code but not likely.

I wanted to make sure that if a listener gets activated that it is able to call the get() method and get the most recent value. To do that I needed to change around the order in which things happened in the setValue.

Basically I needed this.Value = newValue; to happen before the updateActions.